переписан способ получения статистики - удалена цикличность обращения к репозиторию

+++ stat-service/src/main/java/explore/with/me/model/ViewStats.java
@@ -9,5 +9,6 @@ public class ViewStats {
 
     String app;
     String uri;
-    int hits;
+    Long hits;
+
 }
